Refactor into separate functions and impl file
