+++ src/software/ai/hl/stp/stp_tactic_assignment_test.cpp
@@ -4,6 +4,7 @@
 
 #include "software/ai/hl/stp/play/test_plays/halt_test_play.h"
 #include "software/ai/hl/stp/stp.h"
+#include "software/ai/hl/stp/stp_impl.h"
 #include "software/ai/hl/stp/tactic/stop_tactic.h"
 #include "software/ai/hl/stp/tactic/test_tactics/goalie_test_tactic.h"
 #include "software/ai/hl/stp/tactic/test_tactics/move_test_tactic.h"
@@ -77,7 +78,7 @@ TEST_F(STPTacticAssignmentTest,
 
     std::vector<std::shared_ptr<Tactic>> tactics = {move_tactic_1, move_tactic_2};
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     EXPECT_TRUE(allTacticsAssigned(tactics));
 }
@@ -99,7 +100,7 @@ TEST_F(STPTacticAssignmentTest,
 
     std::vector<std::shared_ptr<Tactic>> tactics = {move_tactic_1, move_tactic_2};
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     EXPECT_TRUE(move_tactic_1->getAssignedRobot().has_value());
     EXPECT_FALSE(move_tactic_2->getAssignedRobot().has_value());
@@ -122,7 +123,7 @@ TEST_F(STPTacticAssignmentTest,
 
     std::vector<std::shared_ptr<Tactic>> tactics = {move_tactic_1};
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     EXPECT_TRUE(move_tactic_1->getAssignedRobot().has_value());
 }
@@ -139,7 +140,7 @@ TEST_F(STPTacticAssignmentTest, test_0_tactics_returned_when_there_are_no_robots
 
     std::vector<std::shared_ptr<Tactic>> tactics = {move_tactic_1};
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     EXPECT_FALSE(move_tactic_1->getAssignedRobot().has_value());
 }
@@ -165,7 +166,7 @@ TEST_F(STPTacticAssignmentTest,
     // way around to move_tactic_2. What we expect is that robot_0 will be assigned to
     // move_tactic_2 and "slide over" to make room for robot_1
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     // move_tactic_1 should be the only Tactic assigned a robot, since stop_tactic_1 is a
     // lower priority than move_tactic_1 so it should be dropped since there's only 1
@@ -189,7 +190,7 @@ TEST_F(STPTacticAssignmentTest, test_assigning_1_tactic_to_1_robot)
 
     std::vector<std::shared_ptr<Tactic>> tactics = {move_tactic_1};
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     ASSERT_TRUE(move_tactic_1->getAssignedRobot().has_value());
     EXPECT_EQ(move_tactic_1->getAssignedRobot().value(), robot_0);
@@ -220,7 +221,7 @@ TEST_F(STPTacticAssignmentTest, test_assigning_2_robots_to_2_tactics_no_overlap)
     // Each robot is close to separate tactic destinations. They should each be trivially
     // assigned to the tactic with the destination closest to their position
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     ASSERT_TRUE(move_tactic_1->getAssignedRobot().has_value());
     ASSERT_TRUE(move_tactic_2->getAssignedRobot().has_value());
@@ -261,7 +262,7 @@ TEST_F(STPTacticAssignmentTest, test_assigning_2_robots_to_2_tactics_with_overla
     // way around to move_tactic_2. What we expect is that robot_0 will be assigned to
     // move_tactic_2 and "slide over" to make room for robot_1
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     ASSERT_TRUE(move_tactic_1->getAssignedRobot().has_value());
     ASSERT_TRUE(move_tactic_2->getAssignedRobot().has_value());
@@ -292,7 +293,7 @@ TEST_F(STPTacticAssignmentTest, test_assigning_3_robots_to_2_tactics)
     // robot_2 should not be assigned since both robot_0 and robot_1 are more optimal
     // to assign to the tactics. robot_2 is too far away
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     ASSERT_TRUE(move_tactic_1->getAssignedRobot().has_value());
     ASSERT_TRUE(move_tactic_2->getAssignedRobot().has_value());
@@ -321,7 +322,7 @@ TEST_F(STPTacticAssignmentTest,
                                                     stop_tactic_3};
 
     // If all costs are equal, the robots and tactics are simply paired in order
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     ASSERT_TRUE(stop_tactic_1->getAssignedRobot().has_value());
     ASSERT_TRUE(stop_tactic_2->getAssignedRobot().has_value());
@@ -356,7 +357,7 @@ TEST_F(STPTacticAssignmentTest,
     std::vector<std::shared_ptr<Tactic>> tactics = {stop_tactic_1, move_tactic_1,
                                                     stop_tactic_2};
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     ASSERT_TRUE(stop_tactic_1->getAssignedRobot().has_value());
     ASSERT_TRUE(move_tactic_1->getAssignedRobot().has_value());
@@ -386,7 +387,7 @@ TEST_F(STPTacticAssignmentTest,
 
     std::vector<std::shared_ptr<Tactic>> tactics = {move_tactic_1};
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     ASSERT_TRUE(move_tactic_1->getAssignedRobot().has_value());
     EXPECT_EQ(move_tactic_1->getAssignedRobot().value(), robot_1);
@@ -411,7 +412,7 @@ TEST_F(STPTacticAssignmentTest,
     auto goalie_tactic_2                         = std::make_shared<GoalieTestTactic>();
     std::vector<std::shared_ptr<Tactic>> tactics = {goalie_tactic_1, goalie_tactic_2};
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     EXPECT_FALSE(goalie_tactic_1->getAssignedRobot().has_value());
     EXPECT_FALSE(goalie_tactic_2->getAssignedRobot().has_value());
@@ -440,7 +441,7 @@ TEST_F(STPTacticAssignmentTest,
     auto goalie_tactic_1                         = std::make_shared<GoalieTestTactic>();
     std::vector<std::shared_ptr<Tactic>> tactics = {goalie_tactic_1};
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     EXPECT_TRUE(allTacticsAssigned(tactics));
     ASSERT_TRUE(goalie_tactic_1->getAssignedRobot().has_value());
@@ -450,7 +451,7 @@ TEST_F(STPTacticAssignmentTest,
     friendly_team.assignGoalie(1);
     world.updateFriendlyTeamState(friendly_team);
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     EXPECT_TRUE(allTacticsAssigned(tactics));
     ASSERT_TRUE(goalie_tactic_1->getAssignedRobot().has_value());
@@ -476,7 +477,7 @@ TEST_F(STPTacticAssignmentTest,
     auto goalie_tactic_2                         = std::make_shared<GoalieTestTactic>();
     std::vector<std::shared_ptr<Tactic>> tactics = {goalie_tactic_1, goalie_tactic_2};
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     ASSERT_TRUE(goalie_tactic_1->getAssignedRobot().has_value());
     EXPECT_FALSE(goalie_tactic_2->getAssignedRobot().has_value());
@@ -507,7 +508,7 @@ TEST_F(STPTacticAssignmentTest,
     std::vector<std::shared_ptr<Tactic>> tactics = {move_tactic_1};
     std::vector<Robot> expected_robots_assigned  = {robot_0, robot_1, robot_2};
 
-    stp.assignRobotsToTactics(world, tactics);
+    assignRobotsToTactics(world, tactics);
 
     // Check each tactic is assigned to the intended robot
     for (unsigned int i = 0; i < tactics.size(); i++)
